[zip] Fix the overlay manager start regression
Turns out that the benchmarks only reliably measure the case
when the loaded archives are already in the page cache, and
completely miss the one where every page read hits a fault.
This actually only happens during the boot, as everywhere else
at least some part of the archive has been loaded earlier for
some other reason.

This change ensures that if libziparchive is going to access
some part of the archive via mmap it would tell the OS to
preload it beforehand.

Profiling shows the time spend in the page fault handler to
drop (almost) to zero, bringing the cold boot time back where
it was for non-mmapped case.

Unfortunately, there's no reliable way to add this to the
benchmark: madvise(DONTNEED) for some reason is often ignored
and the pages remain in the cache (often, but not always, making
it even worse). The only good test as of now is the
OverlayManagerService startup metric.

+ fix a wrong cast of the file length, where it was getting
  sliced before the value has been compared

Test: build + unit tests + benchmarks + boot
Bug: 260781425
Change-Id: I1616d77e7209da2fc6ecb9193bf89106df6ef0c6